Use saner names for the pin attributes
I think that `d_or_c` is as confusing as `dc`, possibly more, when the parameter is still called `dc`. If we wanted to be hyper-correct and stick to Python's style, we should call it `data_or_command`, but then for consistency we would also need things like `master_in_slave_out` and so on, and I think this is an overkill. Since the pin is usually labelled `dc` on the physical modules anyways, I think it's perfectly fine to use that — it's also consistent with the usage in Arduino and other drivers. I added `_pin` so make the linter shut up and also to clarify that this is indeed a pin. I also changed the name of `res` to `reset_pin` to be consistent (the label for that pin on the modules is sometimes `rs` and sometimes `rst` so keeping the shortcut is not helpful).